add: split to_rust variable_declaration
