+++ src/to_ts_rust/to_rust/mod.rs
@@ -1,2 +1,3 @@
 pub mod to_rust;
 mod common_methods;
+mod variable_declaration;
