+++ src/to_ts_rust/to_rust/variable_declaration/mod.rs
@@ -1,0 +1,1 @@
+pub mod variable_declaration;
